activity_login.xml and activity_sign_up.xml got a bunch of visual changes, rearranging and the logo was made and added.
colors.xml got the color scheme added in.

edit_background.xml got added in to change the colors and shapes on the buttons since that's how the video did it.

build.gradle had minSDK and targetSDK values changed to work on my phone so if that ends up causing problems this is where it happened. They were changed to my phones SDK version which is 31 as opposed to the 33 used in the experimental API emulator.